Further optimization of DBID layer and heaps. Fixes a performance regressions (and speeds up some more) due to odd optimization problem of hotspot JIT. Enhance heap by adding a "replaceTopElement" option [closes ticket:96] This is not the best solution, but it is the fastest (and comparably reasonable) after days of benchmarking.

+++ src/de/lmu/ifi/dbs/elki/database/datastore/memory/ArrayDoubleStore.java
@@ -77,7 +77,7 @@ public class ArrayDoubleStore implements WritableDoubleDataStore {
   @Deprecated
   public Double get(DBIDRef id) {
     try {
-      return data[idmap.mapDBIDToOffset(id)];
+      return Double.valueOf(data[idmap.mapDBIDToOffset(id)]);
     }
     catch(ArrayIndexOutOfBoundsException e) {
       return null;
@@ -89,8 +89,8 @@ public class ArrayDoubleStore implements WritableDoubleDataStore {
   public Double put(DBIDRef id, Double value) {
     final int off = idmap.mapDBIDToOffset(id);
     double ret = data[off];
-    data[off] = value;
-    return ret;
+    data[off] = value.doubleValue();
+    return Double.valueOf(ret);
   }
 
   @Override
